pyproject.toml
[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=43.0.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["theseus"]

[project]
name = "theseus"
version = "1.5.0"
description = "A general template for various Deep Learning tasks. Strongly relies on Pytorch"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["pytorch", "template", "deep learning"]
authors = [
  {name = "Minh-Khoi Pham", email = "minhkhoi.pham@adaptcentre.ie" },
  {name = "Hoang-Lan Nguyen", email = "nhlan091000@gmail.com"}
]
maintainers = [
  {name = "Minh-Khoi Pham", email = "minhkhoi.pham@adaptcentre.ie" }
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Deep Learning :: Template",
  "License :: MIT License",
  "Programming Language :: Python :: 3",
]
dependencies = [
    "seaborn>=0.12.0",
    "plotly>=5.10.0",
    "matplotlib>=3.4.3",
    "webcolors>=1.12",
    "wandb>=0.13.3",
    "tensorboard>=2.10.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.1",
    "gdown>=4.5.1",
    "tabulate>=0.8.10",
    "loguru>=0.6.0",
    "kaleido>=0.2.1",
    "optuna>=3.0.5",
    "deepdiff>=6.2.3",
    "hydra-core>=1.3.2",
    "lightning>=2.0.0"
]

[project.optional-dependencies]
dev=[
    "pytest>=7.2.0",
    "pytest-order>=1.0.1",
    "pre-commit>=2.21.0",
]
cv = [
    "albumentations>=1.1.0",
    "omegaconf>=2.2.3"
]
cv_classification = [
    "timm",
    "scikit-plot",
    "grad-cam>=1.4.5"
]
cv_semantic = [
    "segmentation-models-pytorch>=0.3.0"
]
cv_detection = [
    "ensemble_boxes>=1.0.9",
    "pycocotools>=2.0.6",
]
nlp = [
  "pyvi>=0.1.1",
  "py_vncorenlp>=0.1.3",
  "spacy>=3.4.2",
  "transformers>=4.25.1",
  "nltk>=3.7",
  "captum>=0.4.1"
]
nlp_retrieval = [
  "sentence_transformers>=2.2.2",
  "rank_bm25>=0.2.2",
  "elasticsearch>=7.17.7"
]
ml = [
  "pandas",
  "pandarallel>=1.6.3",
  "numpy",
  "scikit-learn>=1.0.0",
  "scipy>=1.7.0",
  "optuna>=3.0.5",
  "psycopg2-binary>=2.9.5",
  "gunicorn>=20.1.0",
  "lightgbm>=3.3.3",
  "xgboost<=1.7.1",
  "catboost",
  "shap>=0.41.0",
  "lime>=0.2.0.1",
  "scikit-plot",
]
all = [
  "theseus[cv,cv_classification,cv_semantic,cv_detection,nlp,nlp_retrieval,ml]",
]

[project.urls]  # Optional
"Homepage" = "https://github.com/kaylode/theseus"

[tool.isort]
profile = "black"
